[docs] Convert all resizable examples to Typescript
