+++ src-docs/src/views/resizable_container/resizable_panel_collapsible_options.tsx
@@ -1,45 +1,38 @@
 import React, { useState } from 'react';
 import {
-  EuiText,
   EuiResizableContainer,
   EuiListGroup,
   EuiListGroupItem,
   EuiPanel,
   EuiTitle,
   EuiSpacer,
+  EuiText,
   EuiPage,
-  useIsWithinBreakpoints,
-} from '../../../../src';
+} from '../../../../src/components';
+// @ts-ignore - faker does not have type declarations
 import { fake } from 'faker';
-
-const texts = [];
-
-for (let i = 0; i < 4; i++) {
-  texts.push(<p>{fake('{{lorem.paragraph}}')}</p>);
-}
 
 export default () => {
   const items = [
     {
       id: 1,
       label: 'First item',
-      text: texts[0],
-      active: true,
+      text: <p>{fake('{{lorem.paragraphs}}')}</p>,
     },
     {
       id: 2,
       label: 'Second item',
-      text: texts[1],
+      text: <p>{fake('{{lorem.paragraphs}}')}</p>,
     },
     {
       id: 3,
       label: 'Third item',
-      text: texts[2],
+      text: <p>{fake('{{lorem.paragraphs}}')}</p>,
     },
     {
       id: 4,
       label: 'Forth item',
-      text: texts[3],
+      text: <p>{fake('{{lorem.paragraphs}}')}</p>,
     },
   ];
 
@@ -52,20 +45,20 @@ export default () => {
       size="s"
     />
   ));
-
-  const isMobile = useIsWithinBreakpoints(['xs', 's']);
-  const style = isMobile ? { height: '100%' } : { minHeight: '100%' };
-
   return (
     <EuiPage paddingSize="none">
-      <EuiResizableContainer
-        direction={isMobile ? 'vertical' : 'horizontal'}
-        style={{ height: '400px' }}
-      >
+      <EuiResizableContainer style={{ height: '320px' }}>
         {(EuiResizablePanel, EuiResizableButton) => (
           <>
             <EuiResizablePanel
-              mode="collapsible"
+              mode={[
+                'collapsible',
+                {
+                  className: 'panel-toggle',
+                  'data-test-subj': 'panel-1-toggle',
+                  position: 'top',
+                },
+              ]}
               initialSize={20}
               minSize="10%"
             >
@@ -74,14 +67,31 @@ export default () => {
 
             <EuiResizableButton />
 
-            <EuiResizablePanel mode="main" initialSize={80} minSize="50px">
-              <EuiPanel paddingSize="l" style={style}>
+            <EuiResizablePanel mode="main" initialSize={60} minSize="20%">
+              <EuiPanel paddingSize="l" style={{ minHeight: '100%' }}>
                 <EuiTitle>
                   <p>{itemSelected.label}</p>
                 </EuiTitle>
                 <EuiSpacer />
                 <EuiText>{itemSelected.text}</EuiText>
               </EuiPanel>
+            </EuiResizablePanel>
+
+            <EuiResizableButton />
+
+            <EuiResizablePanel
+              mode={[
+                'collapsible',
+                {
+                  className: 'panel-toggle',
+                  'data-test-subj': 'panel-3-toggle',
+                  position: 'bottom',
+                },
+              ]}
+              initialSize={20}
+              minSize="10%"
+            >
+              <EuiListGroup flush>{itemElements}</EuiListGroup>
             </EuiResizablePanel>
           </>
         )}
