+++ src-docs/src/views/resizable_container/resizable_panels.tsx
